refactor(data_operations): share partition_by, source-column guards, test helper
Extend the dedup pass with the byte-identical clusters from #259:

- extract_partition_by(): one free function in data_operations/base.py,
  replacing the four identical _extract_partition_by copies (ffill, ema,
  sessionization, resample bases).
- assert_source_col_present(): one guard per backend, each in a
  backend-specific helper module so pandas/polars/duckdb stay
  import-isolated (pandas_helpers, new polars_helpers, new duckdb_helpers).
  The per-op _assert_source_column_present overrides now delegate.
- _assert_float_list_with_nulls(): moved into DataOpsTestBase, removing
  the identical copies from the ffill and ema test bases.

No behavior change.

Refs #259.
