+++ mloda/community/feature_groups/data_operations/row_changing/resample/base.py
@@ -46,6 +46,8 @@ from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 from mloda.core.abstract_plugins.components.options import Options
 from mloda.provider import DefaultOptionKeys, FeatureGroup
 
+from mloda.community.feature_groups.data_operations.base import extract_partition_by
+
 # Order-independent aggregations supported in v1. Order-dependent aggregations
 # (e.g. ``first`` / ``last``) and ``median`` are deliberately excluded.
 RESAMPLE_AGGS: dict[str, str] = {
@@ -210,10 +212,7 @@ class ResampleFeatureGroup(FeatureChainParserMixin, FeatureGroup):
     @classmethod
     def _extract_partition_by(cls, feature: Feature) -> list[str]:
         """Return ``partition_by`` as a list (defaulting to ``[]`` when absent)."""
-        partition_by = feature.options.get(cls.PARTITION_BY)
-        if partition_by is None:
-            return []
-        return list(partition_by)
+        return extract_partition_by(feature, cls.PARTITION_BY)
 
     @classmethod
     def _extract_time_column(cls, feature: Feature) -> str:
