+++ mloda/community/feature_groups/data_operations/row_preserving/ema/base.py
@@ -50,6 +50,8 @@ from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 from mloda.core.abstract_plugins.components.options import Options
 from mloda.provider import DefaultOptionKeys, FeatureGroup
 
+from mloda.community.feature_groups.data_operations.base import extract_partition_by
+
 
 class EmaFeatureGroup(FeatureChainParserMixin, FeatureGroup):
     """Base class for exponential-moving-average operations that preserve row count."""
@@ -139,10 +141,7 @@ class EmaFeatureGroup(FeatureChainParserMixin, FeatureGroup):
     @classmethod
     def _extract_partition_by(cls, feature: Feature) -> list[str]:
         """Return ``partition_by`` as a list (defaulting to ``[]`` when absent)."""
-        partition_by = feature.options.get(cls.PARTITION_BY)
-        if partition_by is None:
-            return []
-        return list(partition_by)
+        return extract_partition_by(feature, cls.PARTITION_BY)
 
     @classmethod
     def _extract_order_by(cls, feature: Feature) -> str:
