+++ mloda/community/feature_groups/data_operations/row_preserving/sessionization/pandas_sessionization.py
@@ -8,6 +8,7 @@ from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.pandas.dataframe import PandasDataFrame
 
 from mloda.community.feature_groups.data_operations.helper_columns import unique_helper_name
+from mloda.community.feature_groups.data_operations.pandas_helpers import assert_source_col_present
 from mloda.community.feature_groups.data_operations.row_preserving.sessionization.base import (
     SessionizationFeatureGroup,
 )
@@ -20,10 +21,7 @@ class PandasSessionization(SessionizationFeatureGroup):
 
     @classmethod
     def _assert_source_column_present(cls, data: pd.DataFrame, order_col: str) -> None:
-        if order_col not in data.columns:
-            raise ValueError(
-                f"Source column {order_col!r} is not present in the pandas DataFrame; available: {list(data.columns)}."
-            )
+        assert_source_col_present(data, order_col)
 
     @classmethod
     def _compute_session(
