+++ mloda/community/feature_groups/data_operations/row_preserving/sessionization/polars_lazy_sessionization.py
@@ -10,6 +10,7 @@ from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.polars.lazy_dataframe import PolarsLazyDataFrame
 
 from mloda.community.feature_groups.data_operations.helper_columns import unique_helper_name
+from mloda.community.feature_groups.data_operations.polars_helpers import assert_source_col_present
 from mloda.community.feature_groups.data_operations.row_preserving.sessionization.base import (
     SessionizationFeatureGroup,
 )
@@ -22,9 +23,7 @@ class PolarsLazySessionization(SessionizationFeatureGroup):
 
     @classmethod
     def _assert_source_column_present(cls, data: pl.LazyFrame, order_col: str) -> None:
-        names = data.collect_schema().names()
-        if order_col not in names:
-            raise ValueError(f"Source column {order_col!r} is not present in the polars frame; available: {names}.")
+        assert_source_col_present(data, order_col)
 
     @classmethod
     def _compute_session(
